backend/requirements.txt
PyJWT==2.10.1
PyYAML==6.0.2
SQLAlchemy==2.0.37
annotated-types==0.7.0
anyio==4.8.0
async-lru==2.0.4
bcrypt==4.2.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
pip==24.3.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
sniffio==1.3.1
starlette==0.45.2
text-unidecode==1.3
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.1
